Progress on fuse.
